move certificate verification functions to cert package and change name of functions for readability
